requirements.txt
docxtpl
fpdf
streamlit
python-docx
docxtpl
docxcompose
pypandoc  # optional PDF conversion